Rename shared to android-shared
