+++ android-shared/.gitignore
