+++ android-shared/src/main/java/org/odk/collect/androidshared/livedata/LiveDataUtils.java
@@ -1,4 +1,4 @@
-package org.odk.collect.shared.livedata;
+package org.odk.collect.androidshared.livedata;
 
 import androidx.lifecycle.LiveData;
 import androidx.lifecycle.MediatorLiveData;
