+++ collect_app/build.gradle
@@ -333,7 +333,7 @@ dependencies {
         exclude group: 'org.hamcrest'
     }
 
-    implementation project(':shared')
+    implementation project(':android-shared')
     implementation project(':nbistubs')
     implementation project(':strings')
     implementation project(':material')
