+++ collect_app/src/test/java/org/odk/collect/android/formentry/FormEntryMenuDelegateTest.java
@@ -17,7 +17,7 @@ import org.odk.collect.android.javarosawrapper.FormController;
 import org.odk.collect.android.preferences.screens.GeneralPreferencesActivity;
 import org.odk.collect.android.utilities.ApplicationConstants;
 import org.odk.collect.audiorecorder.recording.AudioRecorder;
-import org.odk.collect.shared.livedata.MutableNonNullLiveData;
+import org.odk.collect.androidshared.livedata.MutableNonNullLiveData;
 import org.robolectric.Robolectric;
 import org.robolectric.annotation.LooperMode;
 import org.robolectric.fakes.RoboMenu;
