+++ settings.gradle
@@ -1,6 +1,6 @@
 include ':forms'
 include ':analytics'
-include ':shared'
+include ':android-shared'
 include ':audiorecorder'
 include ':testshared'
 include ':audioclips'
